描画関係をUVMeshから分離
